requirements.txt
fastapi
uvicorn
sqlalchemy
sentence-transformers
python-multipart
python-docx
PyPDF2
scikit-learn
numpy